pyproject.toml
[project]
name = "smartclinic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "phamkhai108", email = "phamkhai8100@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "elasticsearch>=8.17.2",
    "fastapi[standard]>=0.115.11",
    "just>=0.8.162",
    "numpy>=2.2.4",
    "openai>=1.72.0",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic-settings>=2.8.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.2",
    "scikit-learn>=1.6.1",
    "uvicorn[standard]>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]
